working so far

+++ scripts/startup.py
@@ -1,12 +1,14 @@
 #!/usr/bin/env python3
 
 import re
-import os
+import os, pwd, grp
 import subprocess
 import sys
 import argparse
+import json
+from collections import Counter
 
-from occ.ConfigEditor import ConfigEditor
+from occ.ConfigEditor import ConfigEditor, Operation
 
 from pathlib import Path
 
@@ -19,6 +21,15 @@ if utility:
 if instance_name == None:
     print("Please specify environment variable 'INSTANCE_NAME'")
     exit(1)
+
+arkmanager_config_folder = Path("/etc/arkmanager")
+arkmanager_config = arkmanager_config_folder / "arkmanager.cfg"
+arkmanager_instance_config = arkmanager_config_folder / "instance.cfg"
+ark_mod_folder = Path("/ark/server/ShooterGame/Content/Mods/")
+logs_folder = Path("/ark/logs/")
+logs_instance_folder = logs_folder / f"{instance_name}.arktools"
+
+os.makedirs(logs_instance_folder, exist_ok=True)
 
 # ex 27000:27000/udp
 port_regex = re.compile(r"(\d+):(\d+)(/((udp)|(tcp)))?$")
@@ -44,21 +55,52 @@ def run(command: str):
     process.wait()
     return process.returncode
 
-parser = argparse.ArgumentParser("Config Editor")
+def get_required_mods(changes):
+    for o in changes:
+        o : Operation
+        if o.key == "ark_GameModIds":
+            return list(map(str.strip, o.value.split(',')))
+    return []
 
-arkmanager_config_folder = Path("/etc/arkmanager")
-arkmanager_config = arkmanager_config_folder / "arkmanager.cfg"
-arkmanager_instance_config = arkmanager_config_folder / "instance.cfg"
+def get_installed_mods():
+    for dir in ark_mod_folder.iterdir():
+        if dir.is_dir():
+            yield dir.name
+
+# from https://stackoverflow.com/a/2699996
+# user https://stackoverflow.com/users/156771/tam%c3%a1s
+# user https://stackoverflow.com/users/60075/craig-mcqueen
+def drop_privileges(uid_name='nobody', gid_name='nogroup'):
+    if os.getuid() != 0:
+        # We're not root so, like, whatever dude
+        return
+
+    # Get the uid/gid from the name
+    passwd = pwd.getpwnam(uid_name)
+    running_gid = grp.getgrnam(gid_name).gr_gid
+
+    # Remove group privileges
+    os.setgroups(os.getgrouplist(passwd.pw_name, passwd.pw_gid))
+
+    # Try setting the new uid/gid
+    os.setgid(running_gid)
+    os.setuid(passwd.pw_uid)
+
+    # Set Home
+    os.environ['HOME'] = passwd.pw_dir
+
+parser = argparse.ArgumentParser("Config Editor")
 
 args = parser.parse_args()
 
 # delete all instance config files
-runf("rm /etc/arkmanager/instances/*")
+if any(Path("/etc/arkmanager/instances/").iterdir()):
+    runf("rm /etc/arkmanager/instances/*")
 
 editor = ConfigEditor()
 
 # debug
-editor.debug = 1
+#editor.debug = 1
 # debug
 #editor.output_file_suffix = ".new.cfg"
 
@@ -68,6 +110,7 @@ changes = editor.parse_changeset_file(arkmanager_config_folder / "arkmanager-cha
 # set basic instance configuration
 changes += editor.manual_setn("bashlike", arkmanager_config, {
     "defaultinstance": instance_name,
+    "logdir": logs_instance_folder,
     f"configfile_{instance_name}": str(arkmanager_instance_config),
 })
 
@@ -90,22 +133,37 @@ changes += editor.parse_env()
 # change configuration file
 editor.change_config(changes)
 
-if os.environ.get('UTILITY') == "update":
-    runf("chown -R steam:steam /ark")
+# ensure /ark and all world saves can be written
+runf("chown -R steam:steam /ark")
 
+if os.environ.get('UTILITY') == "update":
     if Path("/ark/server/ShooterGame").exists() and Path("/ark/server/version.txt").exists():
         runf("arkmanager installmods")
         os.execlp("arkmanager", "arkmanager", "update","--update-mods")
     else:
+        # install game
         runf("arkmanager install --verbose")
+
+        # backup default server config files
+        runf("tar cfz /ark/defaultconfig.tgz /ark/server/ShooterGame/SavedArk/Config/LinuxServer/*")
+
+        # install required mods
         os.execlp("arkmanager", "arkmanager", "installmods")
-elif os.environ.get('UTILITY', "no") != "no":
+elif os.environ.get('UTILITY', "no") == "no":
     # check if game is installed and fail if not
     if not Path("/ark/server/version.txt").exists():
         print("ARK Survival Evolved server files are not installed. Please run this container with 'update' command.")
         exit(1)
+    
+    # check mods
+    mods = get_required_mods(changes)
+    installed_mods = list(get_installed_mods())
+    if not set(mods) <= set(installed_mods):
+        run(f"echo Instance '{instance_name}' has uninstalled mods.")
+        runf("arkmanager installmods --verbose")
 
-    instance_name = os.environ.get('INSTANCE_NAME')
+    # avoid runuser process
+    drop_privileges("steam","steam")
 
     # start game server and replace python process
-    os.execlp("arkmanager", "arkmanager", "run", instance_name)
+    os.execlp("arkmanager", "arkmanager", "run", f"@{instance_name}")
